+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* 🎉 feat(slices-unique): unique function added
 * ✅ test(slices-tail): tail test cases and benchmarks added
 * 🎉 feat(slices-tail): tail function added
 * 🐛 fix(slices-same): same function updated so it can compare struct unexported values too